extract ProcessAndExportObjects + begin export progress reporting
